Add suppport for the OpenGL GL_MESA_ycbcr_texture extension, use it under certain circumstances for OpenGL video rendering and move software YUV to RGB conversion into OpenGLContext from OpenGLVideo.
This extension provides hardware colorspace conversion for YUYV frame formats and is available on Intel chipsets. Using it on older integrated Intel GPUs provides playback performance comparable to XVideo whereas previously OpenGL playback was unuseable. Other OpenGL features (OSD, deinterlacing, bicubic) are not available when using ycbcr textures.

libavcodec is used to convert the frame, which does not implement the chroma upsampling fix for interlaced frames. It is, however,unlikely to provide a good trade-off of performance and quality on those systems where it will be used and such systems are most likely to use software bob deinterlacing, which does not require the chroma fix.




git-svn-id: http://svn.mythtv.org/svn/trunk@21429 7dbf422c-18fa-0310-86e9-fd20926502f2