Fix a bug in the way keypress events are handled resulting in jumppoints being activated multiple times for each key press.
The problem is MythMainWindow::TranslateKeypress() can be called several times
for each keypress event resulting in the jumppoint code to be executed each time.
Things are made worse by the conversion to MythUI because the keypress event
can be passed around between MythScreenTypes and MythUITypes and their base
classes until one of them handles the keypress.

The fix involves changing MythMainWindow::TranslateKeypress() to return true if
it handled the event (the keypress is a jump point) or false if the caller
should continue to process the event. All the other changes are just to ensure
everywhere TranslateKeypress() is used the caller will ignore the keypress
event if it has already been processed by TranslateKeypress().

git-svn-id: http://svn.mythtv.org/svn/trunk@21440 7dbf422c-18fa-0310-86e9-fd20926502f2