Refs #6516. One of several patches for the myth socket timeout problem.
This one is the least important, it just handles ANN error conditions and some timeouts better. However, it is the only one of the patches that changes the protocol slightly, so I'm applying this first.

The protocol change is simply that ANN handling returns "ERROR" followed by the reason for the error instead of nothing for certain error conditions.

Because of the protocol change all backends and frontends must be recompiled.



git-svn-id: http://svn.mythtv.org/svn/trunk@21445 7dbf422c-18fa-0310-86e9-fd20926502f2