Add a new 'commercial' flagger "Pre & Post Roll" that is
DESIGNED TO FLAG -=-=-=ONLY=-=-=- THE PRE/POST ROLL PARTS

of a recording so that they can be easily skipped.  This is meant to be
used on channels without commercials such as movie channels that have
their own 'commercials' prior to and following a movie.  This flagger
will attempt to highlight those areas so that they can be easily skipped
during playback.

Closes #6889 using patch by Paul ?Turpin?.



git-svn-id: http://svn.mythtv.org/svn/trunk@21446 7dbf422c-18fa-0310-86e9-fd20926502f2