Refs #6516. This consolidates command socket reconnection code and fixes a couple popups which didn't work correctly after the mythui port of the menu screens.
There is also a hidden option to increase reconnect tries, but I don't know what an optimal value is for this yet so I've left it at the old default for now.

The consolidation allows us to handle an unresponive server consistently. The popup for a version mismatch had not been ported to MythUI yet, and the popup for a general unsuccessful connect didn't pop up when called because we called it before the event thread had started.

git-svn-id: http://svn.mythtv.org/svn/trunk@21505 7dbf422c-18fa-0310-86e9-fd20926502f2