Fix an issue with the debugging messages in DetectLetterbox.
The YV12 format detected message would only ever print if detection was on
when we passed frame #1.  This message is now printed no matter when you turn
on the letterbox detection.

The unsupported frame format message would potentially print out for each
frame if we were looking at a video with non-YV12 data.  This patch suppresses
subsequent messages after the first so we don't spam the logs 25+ times
per second if we hit this situation.



git-svn-id: http://svn.mythtv.org/svn/trunk@21520 7dbf422c-18fa-0310-86e9-fd20926502f2