Fix the XVideo hue range for NVIDIA drivers.
The default range of the hue setting has been changed in more recent drivers, which broke the existing 'hue base' assumption. Instead of trying to guess the appropriate adjustment from the driver version etc, we use the XV_SET_DEFAULTS attribute to restore the xvideo attributes to their defaults and calculate the appropriate range offset from that. The behaviour is unchanged if the XV_SET_DEFAULTS attribute is not available.

We do however need to save the existing attribute values and later restore them. This has the added bonus/side-effect of restoring the xvideo port attributes to the state in which MythTV found them (possibly with the exception of XV_COLORKEY and XV_AUTOPAINT_COLORKEY).

Technically we should probably revisit the offset assumptions for all xvideo attributes in 0.23 and rebase the hue setting generally - as most implementations now seem to use a default of 50% of the range rather than the 0% we assume.

Closes #6849. Closes #6778.


git-svn-id: http://svn.mythtv.org/svn/trunk@21522 7dbf422c-18fa-0310-86e9-fd20926502f2