More fixes and some minor safety enhancements for the Previously Recorded screen.
Added a confirmation popup before deleting all entries for a title.
This action can't be undone easily, so it shouldn't be as easy to
unintentionally invoke.  Added a confirmation for the single episode
delete action to be consistent.

Update the list immediately after deleting entries.  It can be
confusing and frustrating to perform a delete and not see it take
effect until the scheduler gets around to sending a change event.
This is how it worked before the UI port.

Bind the DELETE action directly to the single episode delete operation
instead of to the multiple option popup.  This is how it worked before
the UI port.

Really restore the current position correctly when the list is
updated.  The previous fix was badly broken.

Don't use the warning state for some recording statuses.  The episodes
have already passed, so there's no need to draw attention to them
anymore.  This is how it worked before the UI port.


git-svn-id: http://svn.mythtv.org/svn/trunk@21651 7dbf422c-18fa-0310-86e9-fd20926502f2