Improved handling of the '-geometry' command line option.
Previously this override was processed manually after the main MythTV context was created and the particular settings involved were not overriden for the session. By parsing the override before the context is created and then using the existing OverrideSettingForSession() method, we ensure the overriden settings are pervasive throughout the frontend from the time of creation and hence behaviour is consistent between using the command line geometry override and the normal route of setting the ui size in Setup->Appearance->Screen settings.

This fixes several playback positioning/sizing issues, as some of the screen size settings are accessed directly in the player code.


git-svn-id: http://svn.mythtv.org/svn/trunk@21671 7dbf422c-18fa-0310-86e9-fd20926502f2