Refs #6969. Applies the dbg patch from a couple days ago.
This gets rid of the sock->UpRef()/DownRef() on broadcast, this is already done in PlaybackSock, and since we UpRef it a few lines earlier, the MythSocket can not be deleated while we're here. It can get disconnected so we add checks to see if it is still connected before doing a writeStringList.

svn commit -m Refs

git-svn-id: http://svn.mythtv.org/svn/trunk@21721 7dbf422c-18fa-0310-86e9-fd20926502f2