Add a RemoteFile::Exists(QString &url) method for checking for the existence of a remote file.  This is much faster than having to open a RemoteFile and then check the filesize.
Backend command is:

QUERY_FILE_EXISTS
filename
storagegroup

Return values are 1 or 0 depending on whether the file is found in
the specified Storage Group.  SG fallback is disabled.  If the file
is found, the backend command returns the filename in the next field
after the '1', but RemoteFile::Exists() does not currently use the
resulting filename.

NOTE: This does bump the API version up since it touches libmythdb,
so you'll have to recompile things.



git-svn-id: http://svn.mythtv.org/svn/trunk@21770 7dbf422c-18fa-0310-86e9-fd20926502f2