This patch is part 1 of 2 to fix seektables for MythVideo files which are stored using the new MythVideo Storage Groups support.
The filemarkup table currently stores filenames will a full
absolute pathname.  These patches convert filemarkup to use a
relative pathname.  This may be relative to a Storage Group
directory or one of MythVideo's VideoStarupDir directories.
Post 0.22, we'll likely be merging filemarkup and recordedmarkup
into a shared table which uses a fileID as the key instead of
a chanid/starttime or filename as the key, so this is a short-term
solution for 0.22.

A new routine StorageGroup::GetRelativePathname() converts
an absolute pathname to a pathname relative to a Storage Group or.
a MythVideo VideoStartupDir.

ProgramInfo now uses StorageGroup::GetRelativePathname() whenever
dealing with the filemarkup table for a MythVideo video.

mythcommflag is also modified to allow passing a myth:// url to
the --video option so that a seektable can be generated for a
file stored in a remote Storage Group directory.

This commit does modify libmyth so the binary version is bumped up
which means you'll need to recompile plugins.



git-svn-id: http://svn.mythtv.org/svn/trunk@21819 7dbf422c-18fa-0310-86e9-fd20926502f2