Fix MythVideo initial database creation.
Removes 2 lines from the first schema upgrade that attempted to add columns to videometadata that were already created in InitializeVideoDatabase().  This may prevent a MythTV 0.11 user from upgrading to current, but the columns are actually necessary and used in InitializeVideoDatabase(), now, and the change will allow new users and users of more-recent versions to upgrade.  Any 0.11 users that may want to upgrade should manually add the missing columns (see [2635]).

Comments the code for schema update 1008, as the MythTV (not MythVideo) schema upgrade 1218, which will already have run, removed the table from which we tried to pull data.

Updates the dbver variable so upgrades continue after schema version 1013.

Fixes #7053.


git-svn-id: http://svn.mythtv.org/svn/trunk@21882 7dbf422c-18fa-0310-86e9-fd20926502f2