A new bug and minor usage change in the MythUIFileBrowser.
When the user selected an actual file, the filename should be
returned by GetResultText().  When the user hit the OK button,
the current directory is returned by GetResultText().  In this
case, there is now way to know what (if any) file the user
wanted to select since they moved focus off the list to hit OK.

When the user does pick a directory, the state of the MFileInfo
returned is currently undefined.  This may change in the future
to hold info about the directory selected.



git-svn-id: http://svn.mythtv.org/svn/trunk@21916 7dbf422c-18fa-0310-86e9-fd20926502f2