Fix the scheduler to not let kNotRecording rules fall through the cracks and act like kSingleRecords.
When previewing the case of deleting an existing recording rule, the
schedule editor saves the updated rule as kNotRecording instead of
deleting the rule.  This change makes the scheduler return the right
results in either case.  It might still be desirable for the schedule
editor to delete the rule, but I'll let stuartm handle that.

Refs #7080.
M    scheduler.cpp


git-svn-id: http://svn.mythtv.org/svn/trunk@21926 7dbf422c-18fa-0310-86e9-fd20926502f2