Fix mythtv-setup's checkStoragePaths() and checkImageStoragePaths() so they only log a DB error if there was a DB error (and not when there are no storage groups defined for the host--in which case the DB error is empty and, therefore, useless).
Also fixes the checkStoragePaths logic such that, when run on the master backend, users are alerted of the problem if there are no storage groups defined on the master backend (which should always have a Default storage Group defined).  Previously, the check could have been fooled if a user configured a special storage group on a non-master backend before configuring the master backend.  I did not go to the extent of checking if the defined storage group is the Default storage group, but this change will improve the accuracy of the check without adding a lot of extra code.


git-svn-id: http://svn.mythtv.org/svn/trunk@21941 7dbf422c-18fa-0310-86e9-fd20926502f2