Fix storage group deletion such that if the group is deleted from the master backend, the whole group is deleted from all hosts.  If the group name is "Default", the directory overrides from all remote hosts are deleted and the local Default group is unchanged.  Users must delete entries individually from the master backend Default group.
This fix allows users to clean up storage group definitions for no-longer-in-use hostnames.

The extra description in the dialog when deleting from remote hosts is short and not pretty due to the lack of space in the dialog, but should get the point across.  After mythtv-setup is ported to MythUI, the UI can be changed to be much more functional and allow better management of all storage group information.


git-svn-id: http://svn.mythtv.org/svn/trunk@21943 7dbf422c-18fa-0310-86e9-fd20926502f2