Fixes some confusing DB schema upgrade messages in the SchemaUpgradeWizard by adding a string that gets printed in the logs to indicate which schema we are checking.  Currently only MythTV, MythVideo, and MythMusic are using the schema wizard, but the log messages could be confusing when the relatively low MythVideo and/or MythMusic version numbers were printed out with the same format message as the higher MythTV schema version number. All messages should now have 'MythTV', 'MythVideo', 'MythMusic' in them.
Modify the schema wizard and DB backup util's BackupDB() methods to return
an enum instead of boolean.  Now we can distinguish whether a backup
was made, failed, skipped for empty DB, disabled, etc..  The schema wizard's
screen is now fixed to show the proper message based on the DB upgrade
status.  Previously we would tell the user a backup was located in an
empty location if they had backups disabled (including Windows users where
backups are hardcoded to be disabled).

Fix a few places where we were manually upgrading the DB schema version
rather than using the performActualUpdate() routine.  This meant that in
a few places, we weren't displaying the 'Upgrading to schema version X'
message for that specific version upgrade.  These messages are now
printed manually in the upgrades that bypass performActualUpdate().



git-svn-id: http://svn.mythtv.org/svn/trunk@21955 7dbf422c-18fa-0310-86e9-fd20926502f2