Refs #6897. Applies a conservative workaround to DVB timeout problem.
The DVB driver bug has apparently been fixed as of Linux 2.6.31, so 2.6.28 through 2.6.30 are the only affected releases. In any case I'm leaving the ticket open because it appears we really want two different behaviours depending on the device. For DVB devices we want to wait a long time recovery, while for HD-PVR devices we want to exit poll so we can reset the device, since it will not auto-recover. In all cases we probably are dealing with real error conditions that result in bits missing from the recording so we should note this and mark the recording as damaged.

git-svn-id: http://svn.mythtv.org/svn/trunk@21976 7dbf422c-18fa-0310-86e9-fd20926502f2