Refs #5318. QTimer fixes.
Fixes an unsafe QTimer delete in httpcomms. I don't know if the QTimer usage is safe here, but this whole class should probably be phased out as it appears to be a hack around the lack of an event thread in individual QThread instances. Which was fixed in Qt4.


Fixes a double delete of the m_updateTimer in guidegrid.

Fixes a Qt4 out-of-create-thread start/stop problem with previewVideoRefreshTimer.

git-svn-id: http://svn.mythtv.org/svn/trunk@21977 7dbf422c-18fa-0310-86e9-fd20926502f2