Show the subtitle in a few menus that ask you to take/confirm an action with respect to a recorded program.
If you are deleting a Star Trek from within a long list of Star Trek's it will now say something like "Star Trek -- The Trouble with Tribbles" instead of just "Star Trek". The little bit of logic there is just to determine if to put the subtitle on a seperate line and whether to truncate it and add elipsis marks.

git-svn-id: http://svn.mythtv.org/svn/trunk@22107 7dbf422c-18fa-0310-86e9-fd20926502f2