Thread off the loading of images in MythUIImage.  The default is to create a thread to load the image unless MythUIImage::Load() is called with an argument of 'false' or the desired image already exists in the in-memory image cache.
This is part of a few patches to speed up loading of
artwork, but also affects other MythUIImage use.

If this causes issues, the default value of 'true' in
MythUIImage's Load(bool allowLoadInBackground = true)
can be set to false to disable background loading.

Again, this will require a 'make clean' since libmythui
is being touched, but I'm waiting to update the binary
version until after one more libmyth related change.



git-svn-id: http://svn.mythtv.org/svn/trunk@22169 7dbf422c-18fa-0310-86e9-fd20926502f2