Fixes #7049.  Checks to see if exiting playback at the end of the recording before setting the bookmark.  For users with "Prompt at end of recording," this change just prevents a bookmark from being set at the point where the prompt is displayed (and, since the end-of-recording prompt does not allow setting bookmarks, the change just makes Myth do the right thing).  For users with the end-of-recording prompt disabled, this effectively prevents the user from setting a bookmark within the last 2s of the recording through the "Action on playback exit" (either "Save position and exit" or one of the "Prompt" values).
The fast-fingered (or really unlucky) user with end-of-recording prompts disabled can still set a bookmark within 2s of the end of recording, or any user can set a bookmark very close to the point where playback will exit by using the SELECT key.  Doing so may make it hard for the user to react in time when resuming playback at the bookmark, but those users can simply restart playback from the beginning (INFO|Play from...|Play from beginning) and clear the bookmark or set it at a more appropriate place.

I may change the approach being used to fix this issue after 0.22 is released wi
th some patches I have, but those patches change more than I want to change during a feature freeze.  Alternatively, the root issue (user is unable to react before playback exits if the bookmark is too close to the end of recording) may be fixed as a result of changes to end-of-recording handling related to #6974.



git-svn-id: http://svn.mythtv.org/svn/trunk@22259 7dbf422c-18fa-0310-86e9-fd20926502f2