Make FreeSans the default OSD Font for new users so artistically-challenged users don't end up with FreeMono (which I've been told isn't pretty).
Before, new users got the first font in the list (ordered alphabetically).  Since the only 3 fonts installed in the default installation of MythTV are FreeMono, FreeSans, and FreeSansBold, the first font was FreeMono.

This does not affect OSD themes that explicitly specify a font (assuming that font file exists in either the MythTV configuration directory or share directory, the theme directory, or is a full path to the font file location).  This also does not change the OSD Font for existing users.  It also does not change the caption/subtitle font.


git-svn-id: http://svn.mythtv.org/svn/trunk@22313 7dbf422c-18fa-0310-86e9-fd20926502f2