try to fix timestamps of delayed subtitles. Refs #6753
this should ensure that the subtiltes have a resonable display time
it does not fix the cause, the subtitle packets are returned very late
from the mpeg ts demuxer. It's a regression of the ffmpeg sync in [20797]
the upstream demuxer behaves in the same way.
I intend to make the demuxing independent of the GetFrame for 0.23
so that we handle improperly interleaved streams better

git-svn-id: http://svn.mythtv.org/svn/trunk@22320 7dbf422c-18fa-0310-86e9-fd20926502f2