Workaround for no video on playback issue with XVideo, XvMC & VDPAU video renderers.
The real bug is that there is no synchronization between the setting of the setting of the m_drawEnabled variable and the thread that uses it. This commit does not introduce such syncronization, but prevents the problem from occurring on fast machines. First it makes the m_drawEnabled variable volatile, so that all threads see the same state of that variable relatively quickly. Second, after setting the variable we wait 50 milliseconds, which should be sufficient for all in progress MythUI repaints to finish.



git-svn-id: http://svn.mythtv.org/svn/trunk@22350 7dbf422c-18fa-0310-86e9-fd20926502f2