Initial support for some of the icons on iMON LCD devices
First round of additions to enable lighting up the speaker,
codec and a few function icons on iMON LCD devices. Not
sure if this works with any other display with icons, but
it doesn't harm an iMON VFD w/o icons...

Works quite well with mythtv and mythvideo, when the internal
player is used, other plugins aren't particularly well fleshed
out, and using an external player for mythvideo only lights up
the MOVIE function icon, but that's to be expected...

Additional patches for supporting lighting more icons and
support for more of the plugins would be appreciated, not sure
if I'll get around to 'em myself anytime soon... This much is
a nice start though (thanks to Markus Schulz for the original
patch).

Refs #5745.



git-svn-id: http://svn.mythtv.org/svn/trunk@22477 7dbf422c-18fa-0310-86e9-fd20926502f2