lcddevice: disentangle libmythdb and libavcodec
Move conversion from specific av codec to a generic codec that
has an icon available on imon lcds into avformatdecoder. This
way, we eliminate the unnecessary dependency on libavcodec in
libmythdb, and the resulting code is actually much smaller and
cleaner. (Thanks be unto Isaac for pointing out how crappily
things were being done, as well as to Janne and Jean-Yves for
the interim build fixage). Still works just as well on my own
iMON LCD.

- Reverts [22497] and [22488]
- Refs #5745


git-svn-id: http://svn.mythtv.org/svn/trunk@22513 7dbf422c-18fa-0310-86e9-fd20926502f2