Backport of [22539] to release-0-22-fixes.
When calculating the frame skip amount for fast-forward/rewind, use an
arbitrary value of 30 as the keyframe distance instead of the bogus
value of 1 that is currently used.  I had hoped to have the decoder
pass the actual keyframe distance to NVP, but that won't work for
recordings with variable keyframe distances.  Using 30 as the keyframe
distance is a compromise that helps some on HD-PVR recordings without
adversely affecting other recordings too much.  Refs #6982.


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-22-fixes@22540 7dbf422c-18fa-0310-86e9-fd20926502f2