Fix an issue affecting the Qt painter not drawing parts of the screen under certain circumstances.
In MythUIType::Draw, we were making the assumption that after this
draw, there was nothing more for us to draw.  This is an invalid
assumption when the drawing area is clipped using the Qt painter.
Instead of resetting the dirty QRegion, subtract out the QRect
that we are being instructed to draw.  If there are no more dirty
regions, then we set m_NeedsRedraw to false so we aren't called
unnecesarily.



git-svn-id: http://svn.mythtv.org/svn/trunk@22551 7dbf422c-18fa-0310-86e9-fd20926502f2