Remove the "Qt Style" setting and always use the Windows style.
The default value was likely to be unusable, and allowing users to change the setting meant themers needed to test their qtlook.txt settings 6 times and still allowed for unusable colors with the "Desktop Style" (which meant whatever Qt style the user had installed/specified for her system), and all Qt-/non-themed-widgets are being replaced with mythui-/themed widgets.  The Windows style was also being forced on Mac OS, so the one value works for all platforms.

This change allows all Terra users to benefit from the new qtlook.txt provided in [22526] and [22533].  Thanks to Yeechang Lee for pointing out how big an effect this had on usability and for suggesting Windows style.  Refs #7364.


git-svn-id: http://svn.mythtv.org/svn/trunk@22569 7dbf422c-18fa-0310-86e9-fd20926502f2