Backport several changes from trunk to make the Watch Recordings screen draw faster to improve the user experience when enterring this screen.
The time it can take to initially draw this screen can vary quite a bit
from user to user because of the speed of the user's system, the number
of total recordings, the number of recordings in the currently selected
Recording Group, etc..  These patches in trunk allow the Watch Recordings
screen to draw almost immediately and then have the data filled in when
we finish loading data from the master.  Previously, the Watch Recordings
screen would not even start drawing until after the list of programs was
loaded from the master backend and all relevant widgets were populated.

The binary api version has been updated, so make clean.

This commit includes backports of the following changesets in trunk:

[22552] [22560] [22571] [22575]



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-22-fixes@22580 7dbf422c-18fa-0310-86e9-fd20926502f2