Foundation work to drastically simplify metadata grabbing in MythVideo.  Rather than 5-7 commands per grabber, this will allow for everything to be done in two queries of the source-- a search, and a grab.  The grabber will return all the values including the image URLs.  We are needlessly doing 5-7 lookups per item, so this will simplify things there too.  I have maintained backwards compatibility for now as no scripts use this new approach, but the goal is for all scripts to do this by .23 and remove the (many, many) extra settings.
Also, add a hook so that the grabber can return an InetRef.  This will allow us to gently tweak the TMDB grabber to return the TMDB number, even when passed an IMDB #, thus making the transition to TMDB numbers smooth.

Convert the alpacine grabber to use the new format, which I will document on the wiki in the coming days.


git-svn-id: http://svn.mythtv.org/svn/trunk@22617 7dbf422c-18fa-0310-86e9-fd20926502f2