Backport of [22618].  Refs #7423.
Fixed a typo which caused ProgFinder to be run instead of
ScheduleEditor when done so from the OSD.

Fixed ScheduleEditor and ViewScheduled to notify TV they are finished
when run from the OSD.

Fixed RecordingRule to start a new rule when trying to edit a
previously deleted rule.  This is how pre-0.22 behaved.  A future
change might present a popup letting the user choose whether or not to
proceed.


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-22-fixes@22619 7dbf422c-18fa-0310-86e9-fd20926502f2