Merge over [22623] from trunk.  References #7381.
Fix a segfault in libavformat that appears to be triggered by an AC3
decoding error.  What the AC3 decoding error is triggered by is still
unknown, but it may be in either our RingBuffer or AVFormatDecoder.
This patch just fixes the segfault while we investigate further into
the AC3 decoding error.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-22-fixes@22624 7dbf422c-18fa-0310-86e9-fd20926502f2