Refs #7406.  Fix channel change in the EPG.  After this commit, SELECT changes the channel and dismisses the EPG when watching LiveTV and the selected show's starttime is less than now + SelChangeRecThreshold.  Users who prefer to always edit a schedule should use the EDIT keybinding or access "Edit Schedule" through the MENU.
In the mythui port of guidegrid, the MENU key was changed to show the menu (as it should).  Unfortunately since prior to this change MENU was used to change the channel and dismiss the EPG, this meant that the only way to change channel and dismiss the EPG when in LiveTV was to enable SelectChangesChannel and use the SELECT key.  Since the use of the MENU key to change channels doesn't fit the UI consistency guidelines, it makes sense to just remove the SelectChangesChannel setting and always use its functionality.

This change also fixes SELECT so that it does not attempt to change the channel if not watching LiveTV (when using the alternate EPG while playing back a recording).  Furthermore, to ensure all functionality is always available in spite of SELECT's dual-usage, it adds "Change Channel" to the menu when watching LiveTV and adds "Edit Recording Status" to the menu for shows that match a recording rule (where "Edit Recording Status" is the function SELECT gives for a show that matches a rule and starts after SelChangeRecThreshold minutes--the "allow you to resolve conflicts or change overrides" in keys.txt).

I will backport this change to 0.22-fixes after 0.22 is released.  Until then users should enable the setting, "Use select to change the channel in the program guide" or use the CHANUPDATE keybinding (by default, 'X') to change the channel and then dismiss the EPG with ESCAPE.


git-svn-id: http://svn.mythtv.org/svn/trunk@22674 7dbf422c-18fa-0310-86e9-fd20926502f2