Fix MythXML's GetRecorded method to return proper programFlags.
In [9270], the bookmark and cutlist fields in mythconverg.recorded were changed to booleans, so the approach for deteriming if a bookmark or cutlist applied to the recording changed.  The initial UPnP commit in [10118] used the old approach for working with bookmark and cutlist and was never updated.  (The UPnP patch was most likely developed against an earlier version of Myth and the change was missed since no one is using the programFlags returned by GetRecorded.)


git-svn-id: http://svn.mythtv.org/svn/trunk@22715 7dbf422c-18fa-0310-86e9-fd20926502f2