Remove a setting that's been disconnected since the Qt4 port and rebalance the columns on the settings screen from which it was removed.
In older versions of MythTV, there was a popup that warned the user when a problem was detected in configuring audio before playback.  When Intel HDA cards (without Master mixer controls) started to be used, MythTV was wrongly warning the user that audio might not work when it was unable to set the Master mixer volume.  In [6605], Daniel added a setting, "AudioNag," which allowed the user to disable this popup.  In [17122], the popup was disabled because it was unsafe.  Since there's no longer a popup for users to disable and since the setting is not hooked to anything and since the entire notification needs redesigned if someone really wants to force a warning (after all, a user wanting audio should be able to hear if there is none), remove the disconnected switch.

Note that no one will get popup warnings about audio configuration.  This settings removal does not have any effect at all other than removing dead code.


git-svn-id: http://svn.mythtv.org/svn/trunk@22743 7dbf422c-18fa-0310-86e9-fd20926502f2