Fixes #6862. ProgramList/RecordingList refactor.
This eliminates a few hundred lines of of duplicated code. There had been some divergance in the code that initialized ProgramInfo lists despite the comments in the code on how to keep them in sync. This attempts to merge all of those but if anyone notices a missing flag or some such thing please open a ticket. This also eliminates some of the Q3PtrList compatibility methods, these were left in as part of the initial Qt4 port, but were not flagged as depreciated and so unfortunately were used in some new code. The ProgramDetail class has put into it's own header and cpp file so it can be moved back into to the appropriate library. Finally, the data structure has been changed from a doubly linked list to a deque for greater speed in most places where it is used. (Uncomment '#define PGLIST_USE_LINKED_LIST' if you want to time the linked list version in some algorithm.)

git-svn-id: http://svn.mythtv.org/svn/trunk@22799 7dbf422c-18fa-0310-86e9-fd20926502f2